fix(security): harden parsers against crafted/hostile images
Robustness fixes so a malformed disk/memory image can't crash or DoS the tool:
- VMFS parse_fd_buf: bound the FS3 metadata reads by md_alignment (OOB panic on a
  tiny md_alignment) and guard file_block_size==0 (div-by-zero); + regression tests.
- Allocations sized from untrusted header fields are now bounded: read_exact_alloc
  (take + read_to_end) for length-prefixed reads (vhd/vhdx/qcow2/vmdk/vmdk_scan),
  capped BAT pre-alloc (vhd/vdi/vhdx), capped vmfs block map, bounded ntfs
  record_size (also fixes a shift/negation panic) and attribute length.
- VBox SSM record sizes read via take + read_to_end (varint length-bomb).
- MappedFile::read_at takes u64 so guest-physical / file offsets aren't truncated
  to usize on 32-bit targets (armv7/arm) for >4 GiB snapshots.
- lsass dpapi: validate GUID length in release (was debug_assert only).
- windows process walk: replace an unreachable!() with a safe break.

Files here also carry incidental rustfmt/clippy changes. Verified: build
(default + all-features), 104 tests, clippy (deny), and a live run against a GOAD
DC disk (18/18 AD hashes + LSA secrets + DPAPI).

+++ src/hyperv/layer.rs
@@ -8,7 +8,7 @@
 use std::fs;
 use std::path::Path;
 
-use crate::error::{VmkatzError, Result};
+use crate::error::{Result, VmkatzError};
 use crate::memory::PhysicalMemory;
 
 /// Hyper-V memory layer: provides physical memory from .bin or raw dump files.
@@ -50,12 +50,14 @@ impl PhysicalMemory for HypervLayer {
             buf.fill(0);
             if phys_addr < self.size {
                 let avail = (self.size - phys_addr) as usize;
-                self.mmap.read_at(phys_addr as usize, &mut buf[..avail])
+                self.mmap
+                    .read_at(phys_addr, &mut buf[..avail])
                     .map_err(|_| VmkatzError::UnmappablePhysical(phys_addr))?;
             }
             return Ok(());
         }
-        self.mmap.read_at(phys_addr as usize, buf)
+        self.mmap
+            .read_at(phys_addr, buf)
             .map_err(|_| VmkatzError::UnmappablePhysical(phys_addr))
     }
